problems with getting data
